Move Python utility functions into separate file
