+++ tomviz/Utilities.h
@@ -157,19 +157,6 @@ void createCameraOrbit(vtkSMSourceProxy* data,
 // target dimension barely fits in the view).
 void setupRenderer(vtkRenderer* renderer, vtkImageSliceMapper* mapper);
 
-/// Check for Python error. Prints error and clears it if an error has occurred.
-/// Return true if an error has occurred, false otherwise.
-bool checkForPythonError();
-
-/// Convert a QString to a Python string
-PyObject* toPyObject(const QString& str);
-
-/// Convert a QVariant object into the appropriate Python type
-PyObject* toPyObject(const QVariant& value);
-
-// Convert a QVariantList into a Python list
-PyObject* toPyObject(const QVariantList& list);
-
 extern double offWhite[3];
 }
